refactor:entity 변경 및 interface 생성 * CommentCard에 masterCard 추가
