+++ src/main/java/com/sooum/core/domain/card/entity/font/Font.java
@@ -1,0 +1,5 @@
+package com.sooum.core.domain.card.entity.font;
+
+public enum Font {
+    DEFAULT, CUSTOM
+}
