+++ src/main/java/com/sooum/core/domain/card/entity/imgtype/ImgType.java
@@ -1,0 +1,5 @@
+package com.sooum.core.domain.card.entity.imgtype;
+
+public enum ImgType {
+    DEFAULT, USER
+}
