--- src/main/java/com/sooum/core/domain/card/font/Font.java
@@ -1,5 +1,0 @@
-package com.sooum.core.domain.card.font;
-
-public enum Font {
-    DEFAULT, CUSTOM
-}
