--- src/main/java/com/sooum/core/domain/card/imgtype/ImgType.java
@@ -1,5 +1,0 @@
-package com.sooum.core.domain.card.imgtype;
-
-public enum ImgType {
-    DEFAULT, USER
-}
