+++ src/main/java/com/sooum/core/domain/card/repository/FeedCardRepository.java
@@ -1,0 +1,7 @@
+package com.sooum.core.domain.card.repository;
+
+import com.sooum.core.domain.card.entity.FeedCard;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface FeedCardRepository extends JpaRepository<FeedCard, Long> {
+}
