+++ src/main/java/com/sooum/core/domain/common/entity/BaseEntity.java
@@ -15,5 +15,5 @@ import java.time.LocalDateTime;
 @EntityListeners(value = AuditingEntityListener.class)
 public abstract class BaseEntity {
     @CreatedDate
-    private LocalDateTime createdTime;
+    private LocalDateTime createdAt;
 }
