--- src/main/java/com/sooum/core/domain/member/devicetype/DeviceType.java
@@ -1,5 +1,0 @@
-package com.sooum.core.domain.member.devicetype;
-
-public enum DeviceType {
-    ANDROID, IOS
-}
