+++ src/main/java/com/sooum/core/domain/member/entity/devicetype/DeviceType.java
@@ -1,0 +1,5 @@
+package com.sooum.core.domain.member.entity.devicetype;
+
+public enum DeviceType {
+    ANDROID, IOS
+}
