+++ src/main/java/com/sooum/core/domain/member/repository/MemberRepository.java
@@ -1,0 +1,7 @@
+package com.sooum.core.domain.member.repository;
+
+import com.sooum.core.domain.member.entity.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+}
