+++ src/main/java/com/sooum/core/domain/member/service/MemberService.java
@@ -1,0 +1,11 @@
+package com.sooum.core.domain.member.service;
+
+import com.sooum.core.domain.member.repository.MemberRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class MemberService {
+    private final MemberRepository memberRepository;
+}
