+++ src/main/java/com/sooum/core/domain/notification/entity/notificationtype/NotificationType.java
@@ -1,0 +1,5 @@
+package com.sooum.core.domain.notification.entity.notificationtype;
+
+public enum NotificationType {
+    LIKE, COMMENT, BLOCKED
+}
