--- src/main/java/com/sooum/core/domain/notification/notificationtype/NotificationType.java
@@ -1,5 +1,0 @@
-package com.sooum.core.domain.notification.notificationtype;
-
-public enum NotificationType {
-    LIKE, COMMENT, BLOCKED
-}
